requirements.txt
boto3>=1.26.0
sagemaker>=2.150.0,<3.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0

